+++ AquaEngine.API/Control/Domain/Model/Commands/IncreaseQuantityCommand.cs
@@ -1,3 +1,3 @@
 ﻿namespace AquaEngine.API.Control.Domain.Model.Commands;
 
-public record IncreaseQuantityCommand(int Quantity, long ProductId);
+public record IncreaseQuantityCommand(int Quantity, int ProductId);
